created weather module and moved funcionality into it. removed comments
